// frontend/src/generated/RenderOutputSettings.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LayoutKernel } from "./LayoutKernel";

export type RenderOutputSettings = { format: 'svg' | 'png', width: number, height: number, seed: number, theme: 'dark' | 'light', kernel: LayoutKernel, };